Move qcut tests
