+++ py-polars/tests/unit/operations/test_cut.py
@@ -101,136 +101,6 @@ def test_cut_include_breaks() -> None:
     assert_frame_equal(df_out, df_expected, categorical_as_str=True)
 
 
-def test_qcut() -> None:
-    # series
-    s = pl.Series("foo", [-2, -1, 0, 1, 2])
-    expected = pl.Series(
-        "foo",
-        [
-            "(-inf, -1]",
-            "(-inf, -1]",
-            "(-1, 1]",
-            "(-1, 1]",
-            "(1, inf]",
-        ],
-        dtype=pl.Categorical,
-    )
-    out = s.qcut([0.25, 0.75])
-    assert_series_equal(out, expected, categorical_as_str=True)
-
-    # dataframe
-    df = pl.DataFrame(s)
-
-    # pre-defined quantile probabilities
-    df_expected = pl.DataFrame({"foo": [-2, -1, 0, 1, 2], "qcut": expected})
-
-    # eager
-    df_out = df.with_columns(pl.col("foo").qcut([0.25, 0.75]).alias("qcut"))
-    assert_frame_equal(df_out, df_expected, categorical_as_str=True)
-
-    # lazy
-    df_out = (
-        df.lazy().with_columns(pl.col("foo").qcut([0.25, 0.75]).alias("qcut")).collect()
-    )
-    assert df_out.schema == {"foo": pl.Int64, "qcut": pl.Categorical}
-    assert_frame_equal(df_out, df_expected, categorical_as_str=True)
-
-
-def test_qcut_with_labels() -> None:
-    # series
-    s = pl.Series("foo", [-2, -1, 0, 1, 2])
-    expected = pl.Series("foo", ["a", "a", "b", "b", "c"], dtype=pl.Categorical)
-    out = s.qcut([0.25, 0.75], labels=["a", "b", "c"])
-    assert_series_equal(out, expected, categorical_as_str=True)
-
-    # dataframe
-    df = pl.DataFrame(s)
-
-    # pre-defined quantile probabilities
-    df_expected = pl.DataFrame({"foo": [-2, -1, 0, 1, 2], "qcut": expected})
-
-    # eager
-    df_out = df.with_columns(
-        pl.col("foo").qcut([0.25, 0.75], labels=["a", "b", "c"]).alias("qcut")
-    )
-    assert_frame_equal(df_out, df_expected, categorical_as_str=True)
-
-    # lazy
-    df_out = (
-        df.lazy()
-        .with_columns(
-            pl.col("foo").qcut([0.25, 0.75], labels=["a", "b", "c"]).alias("qcut")
-        )
-        .collect()
-    )
-    assert df_out.schema == {"foo": pl.Int64, "qcut": pl.Categorical}
-    assert_frame_equal(df_out, df_expected, categorical_as_str=True)
-
-    # uniform quantile probabilities
-    df_expected = pl.DataFrame(
-        {
-            "foo": [-2, -1, 0, 1, 2],
-            "qcut": pl.Series(
-                ["low", "low", "high", "high", "high"], dtype=pl.Categorical
-            ),
-        }
-    )
-    # eager
-    df_out = df.with_columns(
-        pl.col("foo").qcut(2, labels=["low", "high"], left_closed=True).alias("qcut")
-    )
-    assert_frame_equal(df_out, df_expected, categorical_as_str=True)
-
-    # lazy
-    df_out = (
-        df.lazy()
-        .with_columns(
-            pl.col("foo")
-            .qcut(2, labels=["low", "high"], left_closed=True)
-            .alias("qcut")
-        )
-        .collect()
-    )
-    assert df_out.schema == {"foo": pl.Int64, "qcut": pl.Categorical}
-    assert_frame_equal(df_out, df_expected, categorical_as_str=True)
-
-
-def test_qcut_include_breaks() -> None:
-    df = pl.DataFrame({"foo": [-2, -1, 0, 1, 2]})
-
-    expected = pl.DataFrame(
-        {
-            "foo": [-2, -1, 0, 1, 2],
-            "brk": [-1.0, -1.0, 1.0, 1.0, inf],
-            "foo_bin": pl.Series(
-                [
-                    "(-inf, -1]",
-                    "(-inf, -1]",
-                    "(-1, 1]",
-                    "(-1, 1]",
-                    "(1, inf]",
-                ],
-                dtype=pl.Categorical,
-            ),
-        }
-    )
-    # eager
-    out = df.with_columns(
-        pl.col("foo").qcut([0.25, 0.75], include_breaks=True).alias("qcut")
-    ).unnest("qcut")
-    assert_frame_equal(out, expected, categorical_as_str=True)
-    # lazy
-    out = (
-        df.lazy()
-        .with_columns(
-            pl.col("foo").qcut([0.25, 0.75], include_breaks=True).alias("qcut")
-        )
-        .unnest("qcut")
-        .collect()
-    )
-    assert out.schema == {"foo": pl.Int64, "brk": pl.Float64, "foo_bin": pl.Categorical}
-    assert_frame_equal(out, expected, categorical_as_str=True)
-
 
 def test_cut_null_values() -> None:
     s = pl.Series([-1.0, None, 1.0, 2.0, None, 8.0, 4.0])
